# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Contacts app",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of contacts app"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://example.com/api/v1"
    }
  ],
  "tags": [
    {
      "name": "Contacts",
      "description": "Operations about contacts."
    },
    {
      "name": "Auth",
      "description": "Auth operations."
    }
  ],
  "paths": {
    "/contacts/{id}": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "Get contact by id",
        "operationId": "GetContactById",
        "description": "Get contact by id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "65ca67e7ae7f10c88b598384"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get contact by id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found contact with id 65ca67e7ae7f10c88b598384!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/contact"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
    "/contacts": {
      "post": {
        "tags": [
          "Contacts"
        ],
        "summary": "Create contact",
        "operationId": "createContact",
        "description": "Create a contact with payload",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "phoneNumber",
                  "contactType"
                ],
                "properties": {
                  "_id": {
                    "description": "The unique identifier of the contact",
                    "type": "string",
                    "example": "66be276eed3fff6781dc8af9"
                  },
                  "name": {
                    "description": "Contact's name",
                    "type": "string",
                    "example": "Ivan"
                  },
                  "phoneNumber": {
                    "description": "Contact's phone",
                    "type": "string",
                    "example": "0000000000"
                  },
                  "email": {
                    "description": "Contact's email",
                    "type": "string",
                    "example": "test@gmail.com"
                  },
                  "isFavourite": {
                    "description": "Indicates whether the contact is marked as a favorite",
                    "type": "boolean",
                    "example": true
                  },
                  "contactType": {
                    "description": "Specifies the type of contact, indicating whether it is for work, home, or personal purposes",
                    "type": "string",
                    "enum": [
                      "work",
                      "home",
                      "personal"
                    ],
                    "example": "work"
                  },
                  "userId": {
                    "description": "The unique identifier of the user who owns this contact",
                    "type": "string",
                    "example": "66b5e364238a937758d83ee7"
                  },
                  "photo": {
                    "description": "Contact's photo",
                    "type": "string",
                    "example": "https://res.cloudinary.com/dzhzzgne2/image/upload/v1723829015/hotz2eqz…"
                  },
                  "createdAt": {
                    "description": "The date and time when the contact was created",
                    "type": "string",
                    "format": "date-time",
                    "example": "2023-01-01T00:00:00.000Z"
                  },
                  "updatedAt": {
                    "description": "The date and time when the contact was last updated",
                    "type": "string",
                    "format": "date-time",
                    "example": "2023-01-01T00:00:00.000Z"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creates a contact",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully created a contact!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/contact"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "contact": {
        "type": "object",
        "required": [
          "_id",
          "name",
          "phoneNumber",
          "contactType",
          "userId"
        ],
        "properties": {
          "_id": {
            "description": "The unique identifier of the contact",
            "type": "string",
            "example": "66be276eed3fff6781dc8af9"
          },
          "name": {
            "description": "Contact's name",
            "type": "string",
            "example": "Ivan"
          },
          "phoneNumber": {
            "description": "Contact's phone",
            "type": "string",
            "example": "0000000000"
          },
          "email": {
            "description": "Contact's email",
            "type": "string",
            "example": "test@gmail.com"
          },
          "isFavourite": {
            "description": "Indicates whether the contact is marked as a favorite",
            "type": "boolean",
            "example": true
          },
          "contactType": {
            "description": "Specifies the type of contact, indicating whether it is for work, home, or personal purposes",
            "type": "string",
            "enum": [
              "work",
              "home",
              "personal"
            ],
            "example": "work"
          },
          "userId": {
            "description": "The unique identifier of the user who owns this contact",
            "type": "string",
            "example": "66b5e364238a937758d83ee7"
          },
          "photo": {
            "description": "Contact's photo",
            "type": "string",
            "example": "https://res.cloudinary.com/dzhzzgne2/image/upload/v1723829015/hotz2eqz…"
          },
          "createdAt": {
            "description": "The date and time when the contact was created",
            "type": "string",
            "format": "date-time",
            "example": "2023-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "description": "The date and time when the contact was last updated",
            "type": "string",
            "format": "date-time",
            "example": "2023-01-01T00:00:00.000Z"
          }
        }
      }
    },
    "responses": {
      "401": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "UnauthorizedError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Access token expired"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "Contact not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "NotFoundError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Contact not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}